Refactor to allow type relations to throw exceptions
